+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 * The MINOR component is used whenever a version has backwards-compatible profile changes. This also indicates that the user can not switch back to a previous MINOR version without using a backup.
 * The PATCH component is used for versions that do not contain profile format changes. Users can freely switch between PATCH versions without risk of data loss.
 
-## [Unreleased]
+## [0.13.4] - 2020-12-26
 
 ### Added
 
@@ -470,7 +470,8 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 
 - Initial public release
 
-[Unreleased]: https://github.com/Snakybo/Clicked/compare/0.13.3...master
+[Unreleased]: https://github.com/Snakybo/Clicked/compare/0.13.4...master
+[0.13.4]: https://github.com/Snakybo/Clicked/releases/tag/0.13.4
 [0.13.3]: https://github.com/Snakybo/Clicked/releases/tag/0.13.3
 [0.13.2]: https://github.com/Snakybo/Clicked/releases/tag/0.13.2
 [0.13.1]: https://github.com/Snakybo/Clicked/releases/tag/0.13.1